Modularize Glossaries folder into components, views and services
