+++ frontend/src/modules/Glossaries/components/index.js
@@ -1,0 +1,6 @@
+export * from './GlossaryAssociations';
+export * from './GlossaryCreateCategoryForm';
+export * from './GlossaryCreateTermForm';
+export * from './GlossaryListItem';
+export * from './GlossaryManagement';
+export * from './GlossaryNodeForm';
