+++ frontend/src/modules/Glossaries/services/deleteCategory.js
