+++ frontend/src/modules/Glossaries/services/deleteGlossary.js
