+++ frontend/src/modules/Glossaries/services/deleteTerm.js
