+++ frontend/src/modules/Glossaries/services/getGlossary.js
