+++ frontend/src/modules/Glossaries/services/getGlossaryTree.js
