+++ frontend/src/modules/Glossaries/services/index.js
@@ -1,0 +1,15 @@
+export * from './approveTermAssociation';
+export * from './createCategory';
+export * from './createGlossary';
+export * from './createTerm';
+export * from './deleteCategory';
+export * from './deleteGlossary';
+export * from './deleteTerm';
+export * from './dismissTermAssociation';
+export * from './getGlossary';
+export * from './getGlossaryTree';
+export * from './listGlossaries';
+export * from './listGlossaryAssociations';
+export * from './updateCategory';
+export * from './updateGlossary';
+export * from './updateTerm';
