+++ frontend/src/modules/Glossaries/services/listGlossaries.js
