+++ frontend/src/modules/Glossaries/services/listGlossaryAssociations.js
