+++ frontend/src/modules/Glossaries/services/updateCategory.js
