+++ frontend/src/services/graphql/Glossary/index.js
@@ -1,21 +1,1 @@
-export * from './approveTermAssociation';
-export * from './createCategory';
-export * from './createGlossary';
-export * from './createTerm';
-export * from './deleteCategory';
-export * from './deleteGlossary';
-export * from './deleteTerm';
-export * from './dismissTermAssociation';
-export * from './getGlossary';
-export * from './getGlossaryTree';
-export * from './getTerm';
-export * from './linkTerm';
-export * from './listAssetLinkedTerms';
-export * from './listGlossaries';
-export * from './listGlossaryAssociations';
-export * from './requestLink';
 export * from './searchGlossary';
-export * from './searchGlossaryHierarchy';
-export * from './updateCategory';
-export * from './updateGlossary';
-export * from './updateTerm';
